refactor(trends): extract source-item subsystem from TrendsService (#752)
Decompose the 1,859-line TrendsService god object into focused, colocated
module services, following the existing modules/ extraction pattern.

New services (each <500 lines, with colocated specs):
- TrendSourceItemsService: live Apify fetch + per-platform normalization +
  fallback synthesis + stored-preview accessors. Collapses the five
  copy-pasted fetch{Platform}SourceItems methods into a single dispatch table.
- TrendSourcePreviewService: preview persistence/precompute/resolution,
  content-feed assembly, the cached getTrendContent endpoint, and discovery
  items.
- TrendQueryService: trend read queries + bootstrap fallback. Unifies the
  near-duplicate findActiveTrends/findLastGoodTrends behind one parameterized
  query helper.
- TrendPrelaunchCorpusService: prelaunch reference-corpus seeding/backfill +
  global corpus stats.

TrendsService becomes a thin orchestrator (1,859 -> 684 lines); Prisma, Cache,
and Apify are no longer injected directly. The shared source-preview limit is
extracted to a single constant to prevent drift.

Public method signatures and runtime behavior are unchanged (verified by the
existing trends.service integration spec); new behavior is covered by colocated
unit specs.

Closes #752
