+++ apps/server/api/src/collections/trends/services/modules/trend-source.constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Default number of source-preview items retained per trend across the trend
+ * source-item subsystem. Shared by the source-items and source-preview services
+ * so the value cannot silently drift between them.
+ */
+export const TREND_SOURCE_PREVIEW_LIMIT = 5;
