+++ apps/server/api/src/collections/trends/services/trends.service.ts
@@ -1,16 +1,9 @@
 import { BrandsService } from '@api/collections/brands/services/brands.service';
 import { CredentialsService } from '@api/collections/credentials/services/credentials.service';
-import {
-  buildPrelaunchReferenceCorpusSeeds,
-  PRELAUNCH_REFERENCE_CORPUS_MINIMUMS,
-  PRELAUNCH_REFERENCE_CORPUS_VERSION,
-  type PrelaunchReferenceCorpusSeed,
-} from '@api/collections/trends/data/prelaunch-reference-corpus.seed';
 import { TrendIdea } from '@api/collections/trends/dto/trend-ideas.dto';
 import { TrendEntity } from '@api/collections/trends/entities/trend.entity';
 import type {
   HistoricalTrendsOptions,
-  TrendContentItem,
   TrendContentResult,
   TrendData,
   TrendDiscoveryItem,
@@ -23,7 +16,6 @@ import type {
   TrendTimelineEntry,
   TrendTurnoverStats,
 } from '@api/collections/trends/interfaces/trend-turnover.interface';
-import type { TrendDocument } from '@api/collections/trends/schemas/trend.schema';
 import type { TrendingHashtagDocument } from '@api/collections/trends/schemas/trending-hashtag.schema';
 import type { TrendingSoundDocument } from '@api/collections/trends/schemas/trending-sound.schema';
 import type { TrendingVideoDocument } from '@api/collections/trends/schemas/trending-video.schema';
@@ -31,159 +23,34 @@ import { TrendAnalysisService } from '@api/collections/trends/services/modules/t
 import { TrendContentIdeasService } from '@api/collections/trends/services/modules/trend-content-ideas.service';
 import { TrendFetchService } from '@api/collections/trends/services/modules/trend-fetch.service';
 import { TrendFilteringService } from '@api/collections/trends/services/modules/trend-filtering.service';
+import {
+  type PrelaunchReferenceCorpusBackfillOptions,
+  type PrelaunchReferenceCorpusBackfillResult,
+  TrendPrelaunchCorpusService,
+} from '@api/collections/trends/services/modules/trend-prelaunch-corpus.service';
+import { TrendQueryService } from '@api/collections/trends/services/modules/trend-query.service';
+import { TrendSourcePreviewService } from '@api/collections/trends/services/modules/trend-source-preview.service';
 import { TrendVideoService } from '@api/collections/trends/services/modules/trend-video.service';
 import { TrendPreferencesService } from '@api/collections/trends/services/trend-preferences.service';
 import { TrendReferenceCorpusService } from '@api/collections/trends/services/trend-reference-corpus.service';
-import { CacheService } from '@api/services/cache/services/cache.service';
-import type {
-  ApifyInstagramPost,
-  ApifyNormalizedTweet,
-  ApifyRedditPost,
-  ApifyTikTokVideo,
-  ApifyYouTubeVideo,
-} from '@api/services/integrations/apify/interfaces/apify.interfaces';
-import { ApifyService } from '@api/services/integrations/apify/services/apify.service';
-import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
 import { Timeframe } from '@genfeedai/enums';
 import { LoggerService } from '@libs/logger/logger.service';
 import { Injectable } from '@nestjs/common';
 
-export interface PrelaunchReferenceCorpusBackfillOptions {
-  dryRun?: boolean;
-  now?: Date;
-}
+export type {
+  PrelaunchReferenceCorpusBackfillOptions,
+  PrelaunchReferenceCorpusBackfillResult,
+} from '@api/collections/trends/services/modules/trend-prelaunch-corpus.service';
 
-export interface PrelaunchReferenceCorpusBackfillResult {
-  createdTrends: number;
-  dryRun: boolean;
-  plannedCreates: number;
-  plannedUpdates: number;
-  referenceSync: {
-    links: number;
-    references: number;
-    snapshots: number;
-  };
-  seedReferences: number;
-  seedTrends: number;
-  updatedTrends: number;
-  version: string;
-}
-
+/**
+ * Orchestrates trend retrieval, access control, and discovery, delegating the
+ * source-item subsystem, preview persistence, read queries, and prelaunch
+ * corpus seeding to focused module services (issue #752).
+ */
 @Injectable()
 export class TrendsService {
-  private readonly CONTENT_CACHE_TTL_SECONDS = 600;
-  private readonly SOURCE_PREVIEW_LIMIT = 5;
-  private readonly CONTENT_FEED_PLATFORMS = new Set([
-    'instagram',
-    'linkedin',
-    'reddit',
-    'tiktok',
-    'twitter',
-    'youtube',
-  ]);
-  private readonly BOOTSTRAP_TRENDS = [
-    {
-      growthRate: 68,
-      mentions: 42_000,
-      metadata: {
-        hashtags: ['#AIAgents', '#WorkflowAutomation'],
-        sampleContent:
-          'Creators are showing how AI agents turn recurring workflows into reusable content systems.',
-        source: 'curated',
-        sourcePreviewState: 'fallback',
-        trendType: 'topic',
-        urls: ['https://genfeed.ai/articles'],
-      },
-      platform: 'twitter',
-      topic: 'AI agent workflow demos',
-      viralityScore: 78,
-    },
-    {
-      growthRate: 55,
-      mentions: 31_000,
-      metadata: {
-        hashtags: ['#CreatorOps', '#ContentSystems'],
-        sampleContent:
-          'Teams are packaging trend research, briefs, reviews, and publishing into repeatable creator ops loops.',
-        source: 'curated',
-        sourcePreviewState: 'fallback',
-        trendType: 'topic',
-        urls: ['https://genfeed.ai/workflows'],
-      },
-      platform: 'linkedin',
-      topic: 'Creator ops playbooks',
-      viralityScore: 72,
-    },
-    {
-      growthRate: 49,
-      mentions: 27_500,
-      metadata: {
-        hashtags: ['#VideoRepurposing', '#ShortFormVideo'],
-        sampleContent:
-          'Short-form creators are remixing long-form clips into hooks, captions, and platform-native edits.',
-        source: 'curated',
-        sourcePreviewState: 'fallback',
-        trendType: 'video',
-        urls: ['https://genfeed.ai/studio'],
-      },
-      platform: 'youtube',
-      topic: 'Clip remix systems',
-      viralityScore: 69,
-    },
-    {
-      growthRate: 61,
-      mentions: 38_500,
-      metadata: {
-        hashtags: ['#CarouselDesign', '#CreatorStrategy'],
-        sampleContent:
-          'Instagram teams are turning dense strategy notes into swipeable carousel lessons and remixable Reel hooks.',
-        source: 'curated',
-        sourcePreviewState: 'fallback',
-        trendType: 'post',
-        urls: ['https://genfeed.ai/studio'],
-      },
-      platform: 'instagram',
-      topic: 'Carousel-to-Reel content systems',
-      viralityScore: 74,
-    },
-    {
-      growthRate: 64,
-      mentions: 44_200,
-      metadata: {
-        hashtags: ['#TikTokTrends', '#AICreators'],
-        sampleContent:
-          'TikTok creators are testing fast before-and-after demos that show a manual content workflow becoming automated.',
-        source: 'curated',
-        sourcePreviewState: 'fallback',
-        trendType: 'video',
-        urls: ['https://genfeed.ai/workflows'],
-      },
-      platform: 'tiktok',
-      topic: 'Automation before-and-after demos',
-      viralityScore: 76,
-    },
-    {
-      growthRate: 42,
-      mentions: 18_700,
-      metadata: {
-        hashtags: ['#CreatorTools', '#SaaS'],
-        sampleContent:
-          'Reddit discussions are comparing lightweight creator stacks for briefs, assets, scheduling, and analytics.',
-        source: 'curated',
-        sourcePreviewState: 'fallback',
-        trendType: 'discussion',
-        urls: ['https://genfeed.ai/articles'],
-      },
-      platform: 'reddit',
-      topic: 'Lean creator stack comparisons',
-      viralityScore: 63,
-    },
-  ] as const;
-
   constructor(
-    private readonly prisma: PrismaService,
     private readonly loggerService: LoggerService,
-    private readonly cacheService: CacheService,
     private readonly brandsService: BrandsService,
     private readonly credentialsService: CredentialsService,
     private readonly trendPreferencesService: TrendPreferencesService,
@@ -193,59 +60,16 @@ export class TrendsService {
     private readonly trendAnalysisService: TrendAnalysisService,
     private readonly trendFilteringService: TrendFilteringService,
     private readonly trendVideoService: TrendVideoService,
-    private readonly apifyService: ApifyService,
+    private readonly trendQueryService: TrendQueryService,
+    private readonly trendSourcePreviewService: TrendSourcePreviewService,
+    private readonly trendPrelaunchCorpusService: TrendPrelaunchCorpusService,
   ) {}
 
-  private toSyncTrendInput(trend: TrendEntity): {
-    id: string;
-    mentions: number;
-    platform: string;
-    sourcePreview: TrendSourceItem[];
-    sourcePreviewState: 'live' | 'fallback' | 'empty';
-    topic: string;
-    viralityScore: number;
-  } {
-    return {
-      id: String(trend.id),
-      mentions: trend.mentions,
-      platform: trend.platform,
-      sourcePreview: this.getStoredTrendSourcePreview(
-        trend,
-        this.SOURCE_PREVIEW_LIMIT,
-      ),
-      sourcePreviewState: this.getStoredTrendSourcePreviewState(
-        trend,
-        this.getStoredTrendSourcePreview(trend, this.SOURCE_PREVIEW_LIMIT),
-      ),
-      topic: trend.topic,
-      viralityScore: trend.viralityScore,
-    };
-  }
-
-  async getGlobalCorpusStats(): Promise<{
+  getGlobalCorpusStats(): Promise<{
     activeTrends: number;
     referenceRecords: number;
   }> {
-    const now = new Date();
-    const [allGlobalTrends, referenceRecords] = await Promise.all([
-      this.prisma.trend.findMany({
-        where: { isDeleted: false, organizationId: null },
-      }),
-      this.trendReferenceCorpusService.countGlobalReferences(),
-    ]);
-    const activeTrends = allGlobalTrends.filter((doc) => {
-      const d = doc.data as unknown as Record<string, unknown>;
-      return (
-        d.isCurrent === true &&
-        d.expiresAt != null &&
-        new Date(d.expiresAt as string) > now
-      );
-    }).length;
-
-    return {
-      activeTrends,
-      referenceRecords,
-    };
+    return this.trendPrelaunchCorpusService.getGlobalCorpusStats();
   }
 
   // ==================== Orchestrator Methods ====================
@@ -291,14 +115,19 @@ export class TrendsService {
       brandId,
       (trend) => this.trendFilteringService.calculateViralityScore(trend),
     );
-    const hydratedTrends = await this.precomputeTrendSourcePreview(trends, {
-      force: true,
-      limit: this.SOURCE_PREVIEW_LIMIT,
-    });
+    const hydratedTrends =
+      await this.trendSourcePreviewService.precomputeTrendSourcePreview(
+        trends,
+        { force: true },
+      );
     await this.trendReferenceCorpusService.syncTrendReferences(
-      hydratedTrends.map((trend) => this.toSyncTrendInput(trend)),
+      hydratedTrends.map((trend) =>
+        this.trendSourcePreviewService.toSyncTrendInput(trend),
+      ),
     );
-    await this.cacheService.invalidateByTags(['trends:content']);
+    await this.trendSourcePreviewService.invalidateContentCache([
+      'trends:content',
+    ]);
     return hydratedTrends;
   }
 
@@ -314,7 +143,7 @@ export class TrendsService {
       allowFetchIfMissing?: boolean;
     } = {},
   ): Promise<TrendEntity[]> {
-    const cachedTrends = await this.findActiveTrends({
+    const cachedTrends = await this.trendQueryService.findActiveTrends({
       brandId: brandId ?? null,
       organizationId: organizationId ?? null,
       platform,
@@ -325,7 +154,7 @@ export class TrendsService {
     }
 
     if (organizationId || brandId) {
-      const globalTrends = await this.findActiveTrends({
+      const globalTrends = await this.trendQueryService.findActiveTrends({
         brandId: null,
         organizationId: null,
         platform,
@@ -338,7 +167,7 @@ export class TrendsService {
       }
     }
 
-    const lastGoodTrends = await this.findLastGoodTrends({
+    const lastGoodTrends = await this.trendQueryService.findLastGoodTrends({
       brandId: brandId ?? null,
       organizationId: organizationId ?? null,
       platform,
@@ -351,11 +180,12 @@ export class TrendsService {
     }
 
     if (organizationId || brandId) {
-      const globalLastGoodTrends = await this.findLastGoodTrends({
-        brandId: null,
-        organizationId: null,
-        platform,
-      });
+      const globalLastGoodTrends =
+        await this.trendQueryService.findLastGoodTrends({
+          brandId: null,
+          organizationId: null,
+          platform,
+        });
       if (globalLastGoodTrends.length > 0) {
         this.loggerService.log(
           'No tenant-scoped last-good trends found, falling back to global last-good trend dataset',
@@ -384,199 +214,7 @@ export class TrendsService {
     this.loggerService.log(
       'Fresh trend fetch returned no data, returning bootstrap trend fallback',
     );
-    return this.getBootstrapTrends(platform);
-  }
-
-  private toTrendEntity(
-    doc: {
-      data: unknown;
-    } & Record<string, unknown>,
-  ): TrendEntity {
-    return new TrendEntity({
-      ...doc,
-      ...(doc.data as Record<string, unknown>),
-    } as unknown as TrendDocument);
-  }
-
-  private async findPrelaunchCorpusDocs(): Promise<
-    Array<{ data: unknown; id: string } & Record<string, unknown>>
-  > {
-    const docs = await this.prisma.trend.findMany({
-      orderBy: { createdAt: 'desc' },
-      take: PRELAUNCH_REFERENCE_CORPUS_MINIMUMS.trends * 2,
-      where: {
-        brandId: null,
-        isDeleted: false,
-        organizationId: null,
-      } as never,
-    });
-
-    return docs.filter((doc) => this.getPrelaunchCorpusKey(doc) != null);
-  }
-
-  private getPrelaunchCorpusKey(doc: { data: unknown }): string | null {
-    const data = doc.data as Record<string, unknown>;
-    const metadata = data.metadata as Record<string, unknown> | undefined;
-    const key = metadata?.prelaunchCorpusKey;
-
-    return typeof key === 'string' && key.length > 0 ? key : null;
-  }
-
-  private buildPrelaunchTrendData(
-    seed: PrelaunchReferenceCorpusSeed,
-    now: Date,
-  ): Record<string, unknown> {
-    const expiresAt = new Date(now.getTime() + 30 * 24 * 60 * 60 * 1000);
-    const sourceUrls = seed.sourcePreview.map((item) => item.sourceUrl);
-    const metadata = {
-      ...seed.metadata,
-      prelaunchCorpus: true,
-      prelaunchCorpusKey: seed.key,
-      source: 'curated',
-      sourcePreviewCache: seed.sourcePreview,
-      sourcePreviewCachedAt: now.toISOString(),
-      sourcePreviewState: 'fallback',
-      sourceSetVersion: PRELAUNCH_REFERENCE_CORPUS_VERSION,
-      trendType: seed.sourcePreview.some((item) => item.contentType === 'video')
-        ? 'video'
-        : 'topic',
-      urls: sourceUrls,
-    };
-
-    return {
-      expiresAt: expiresAt.toISOString(),
-      growthRate: seed.growthRate,
-      isCurrent: true,
-      isDeleted: false,
-      mentions: seed.mentions,
-      metadata,
-      platform: seed.platform,
-      requiresAuth: false,
-      topic: seed.topic,
-      viralityScore: seed.viralityScore,
-    };
-  }
-
-  private async findActiveTrends(filter: {
-    organizationId: string | null;
-    brandId: string | null;
-    platform?: string;
-  }): Promise<TrendEntity[]> {
-    const now = new Date();
-    const docs = await this.prisma.trend.findMany({
-      orderBy: { createdAt: 'desc' },
-      take: 200,
-      where: {
-        brandId: filter.brandId,
-        isDeleted: false,
-        organizationId: filter.organizationId,
-      },
-    });
-
-    return docs
-      .filter((doc) => {
-        const d = doc.data as unknown as Record<string, unknown>;
-        if (d.isCurrent !== true) return false;
-        if (d.expiresAt && new Date(d.expiresAt as string) <= now) return false;
-        if (filter.platform && d.platform !== filter.platform) return false;
-        return true;
-      })
-      .sort((a, b) => {
-        const ad = a.data as unknown as Record<string, number>;
-        const bd = b.data as unknown as Record<string, number>;
-        return (bd.viralityScore ?? 0) - (ad.viralityScore ?? 0);
-      })
-      .slice(0, 50)
-      .map((doc) => this.toTrendEntity(doc));
-  }
-
-  private async findLastGoodTrends(filter: {
-    organizationId: string | null;
-    brandId: string | null;
-    platform?: string;
-  }): Promise<TrendEntity[]> {
-    const docs = await this.prisma.trend.findMany({
-      orderBy: { createdAt: 'desc' },
-      take: 200,
-      where: {
-        brandId: filter.brandId,
-        isDeleted: false,
-        organizationId: filter.organizationId,
-      },
-    });
-
-    return docs
-      .filter((doc) => {
-        const d = doc.data as unknown as Record<string, unknown>;
-        if (filter.platform && d.platform !== filter.platform) return false;
-        return true;
-      })
-      .sort((a, b) => {
-        const ad = a.data as unknown as Record<string, number>;
-        const bd = b.data as unknown as Record<string, number>;
-        return (
-          (bd.viralityScore ?? 0) - (ad.viralityScore ?? 0) ||
-          new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-        );
-      })
-      .slice(0, 50)
-      .map((doc) => this.toTrendEntity(doc));
-  }
-
-  private getBootstrapTrends(platform?: string): TrendEntity[] {
-    const now = new Date();
-    const expiresAt = new Date(now.getTime() + 6 * 60 * 60 * 1000);
-
-    return this.BOOTSTRAP_TRENDS.filter(
-      (trend) => !platform || trend.platform === platform,
-    ).map((trend, index) => {
-      const id = `bootstrap-trend-${trend.platform}-${index + 1}`;
-      const metadata = {
-        ...trend.metadata,
-        sourcePreviewCache: [
-          {
-            contentType:
-              trend.platform === 'twitter'
-                ? 'tweet'
-                : trend.metadata.trendType === 'video'
-                  ? 'video'
-                  : 'post',
-            id: `${id}-fallback-1`,
-            platform: trend.platform,
-            sourceUrl: trend.metadata.urls[0],
-            text: trend.metadata.sampleContent,
-            title: trend.topic,
-          },
-        ],
-        sourcePreviewCachedAt: now.toISOString(),
-      };
-
-      return new TrendEntity({
-        brandId: null,
-        createdAt: now,
-        data: {
-          ...trend,
-          expiresAt,
-          isCurrent: true,
-          isDeleted: false,
-          metadata,
-          requiresAuth: false,
-        },
-        expiresAt,
-        growthRate: trend.growthRate,
-        id,
-        isCurrent: true,
-        isDeleted: false,
-        mentions: trend.mentions,
-        metadata,
-        organizationId: null,
-        platform: trend.platform,
-        requiresAuth: false,
-        topic: trend.topic,
-        updatedAt: now,
-        viralityScore: trend.viralityScore,
-      } as never);
-    });
+    return this.trendQueryService.getBootstrapTrends(platform);
   }
 
   /**
@@ -647,7 +285,7 @@ export class TrendsService {
     // through this wrapper and still need a curated set to render, so restore the
     // bootstrap fallback here when nothing is cached.
     if (trends.length === 0) {
-      trends = this.getBootstrapTrends(platform);
+      trends = this.trendQueryService.getBootstrapTrends(platform);
     }
 
     // Filter by brand description if brandId provided
@@ -734,7 +372,10 @@ export class TrendsService {
 
     const trends = await Promise.all(
       result.trends.map(async (trend) =>
-        this.buildTrendDiscoveryItem(trend, organizationId),
+        this.trendSourcePreviewService.buildTrendDiscoveryItem(
+          trend,
+          organizationId,
+        ),
       ),
     );
 
@@ -787,122 +428,35 @@ export class TrendsService {
 
   async precomputeGlobalTrendSourcePreview(): Promise<{ processed: number }> {
     const trends = await this.getTrends();
-    const hydratedTrends = await this.precomputeTrendSourcePreview(trends, {
-      force: true,
-      limit: this.SOURCE_PREVIEW_LIMIT,
-    });
+    const hydratedTrends =
+      await this.trendSourcePreviewService.precomputeTrendSourcePreview(
+        trends,
+        { force: true },
+      );
     await this.trendReferenceCorpusService.syncTrendReferences(
-      hydratedTrends.map((trend) => this.toSyncTrendInput(trend)),
+      hydratedTrends.map((trend) =>
+        this.trendSourcePreviewService.toSyncTrendInput(trend),
+      ),
     );
     const processed = hydratedTrends.length;
 
-    await this.cacheService.invalidateByTags(['trends:content']);
+    await this.trendSourcePreviewService.invalidateContentCache([
+      'trends:content',
+    ]);
 
     return { processed };
   }
 
-  async backfillPrelaunchReferenceCorpus(
-    options: PrelaunchReferenceCorpusBackfillOptions = {},
-  ): Promise<PrelaunchReferenceCorpusBackfillResult> {
-    const now = options.now ?? new Date();
-    const dryRun = options.dryRun ?? true;
-    const seeds = buildPrelaunchReferenceCorpusSeeds(now);
-    const seedReferences = seeds.reduce(
-      (total, seed) => total + seed.sourcePreview.length,
-      0,
+  backfillPrelaunchReferenceCorpus(
+    options: Parameters<
+      TrendPrelaunchCorpusService['backfillPrelaunchReferenceCorpus']
+    >[0] = {},
+  ): ReturnType<
+    TrendPrelaunchCorpusService['backfillPrelaunchReferenceCorpus']
+  > {
+    return this.trendPrelaunchCorpusService.backfillPrelaunchReferenceCorpus(
+      options,
     );
-
-    const existingDocs = await this.findPrelaunchCorpusDocs();
-    const existingByKey = new Map(
-      existingDocs.flatMap((doc) => {
-        const key = this.getPrelaunchCorpusKey(doc);
-        return key ? [[key, doc] as const] : [];
-      }),
-    );
-    const plannedCreates = seeds.filter(
-      (seed) => !existingByKey.has(seed.key),
-    ).length;
-    const plannedUpdates = seeds.length - plannedCreates;
-
-    if (dryRun) {
-      return {
-        createdTrends: 0,
-        dryRun,
-        plannedCreates,
-        plannedUpdates,
-        referenceSync: { links: 0, references: 0, snapshots: 0 },
-        seedReferences,
-        seedTrends: seeds.length,
-        updatedTrends: 0,
-        version: PRELAUNCH_REFERENCE_CORPUS_VERSION,
-      };
-    }
-
-    const syncedTrends: TrendEntity[] = [];
-    let createdTrends = 0;
-    let updatedTrends = 0;
-
-    for (const seed of seeds) {
-      const trendData = this.buildPrelaunchTrendData(seed, now);
-      const existing = existingByKey.get(seed.key);
-
-      if (existing) {
-        const updated = await this.prisma.trend.update({
-          data: {
-            data: trendData,
-            isDeleted: false,
-          } as never,
-          where: { id: existing.id },
-        });
-        syncedTrends.push(this.toTrendEntity(updated));
-        updatedTrends += 1;
-        continue;
-      }
-
-      const created = await this.prisma.trend.create({
-        data: {
-          brandId: null,
-          data: trendData,
-          isDeleted: false,
-          organizationId: null,
-        } as never,
-      });
-      syncedTrends.push(this.toTrendEntity(created));
-      createdTrends += 1;
-    }
-
-    const referenceSync =
-      await this.trendReferenceCorpusService.syncTrendReferences(
-        syncedTrends.map((trend) => this.toSyncTrendInput(trend)),
-      );
-
-    await this.cacheService.invalidateByTags(['trends', 'trends:content']);
-
-    if (
-      seeds.length < PRELAUNCH_REFERENCE_CORPUS_MINIMUMS.trends ||
-      seedReferences < PRELAUNCH_REFERENCE_CORPUS_MINIMUMS.sourceReferences
-    ) {
-      this.loggerService.warn(
-        'Prelaunch reference corpus seed is below floor',
-        {
-          minimums: PRELAUNCH_REFERENCE_CORPUS_MINIMUMS,
-          seedReferences,
-          seedTrends: seeds.length,
-        },
-      );
-    }
-
-    return {
-      createdTrends,
-      dryRun,
-      plannedCreates,
-      plannedUpdates,
-      referenceSync,
-      seedReferences,
-      seedTrends: seeds.length,
-      updatedTrends,
-      version: PRELAUNCH_REFERENCE_CORPUS_VERSION,
-    };
   }
 
   async getReferenceCorpus(
@@ -961,42 +515,16 @@ export class TrendsService {
     );
   }
 
-  // ==================== Delegated: Filtering ====================
+  // ==================== Delegated: Read Queries ====================
 
   /**
    * Get a single trend by ID
    */
-  async getTrendById(
+  getTrendById(
     trendId: string,
     organizationId?: string,
   ): Promise<TrendEntity | null> {
-    const doc = await this.prisma.trend.findFirst({
-      where: {
-        id: trendId,
-        isDeleted: false,
-        ...(organizationId
-          ? { OR: [{ organizationId }, { organizationId: null }] }
-          : {}),
-      },
-    });
-
-    if (!doc) {
-      return null;
-    }
-
-    // If organizationId provided, trend must belong to that org or be global
-    if (organizationId) {
-      const docOrgId = (doc as unknown as Record<string, unknown>)
-        .organizationId;
-      if (docOrgId !== organizationId && docOrgId !== null) {
-        return null;
-      }
-    }
-
-    return new TrendEntity({
-      ...doc,
-      ...(doc.data as Record<string, unknown>),
-    } as unknown as TrendDocument);
+    return this.trendQueryService.getTrendById(trendId, organizationId);
   }
 
   async getTrendSourceItems(
@@ -1004,32 +532,18 @@ export class TrendsService {
     organizationId?: string,
     limit: number = 5,
   ): Promise<TrendSourceItem[]> {
-    const trend = await this.getTrendById(trendId, organizationId);
+    const trend = await this.trendQueryService.getTrendById(
+      trendId,
+      organizationId,
+    );
     if (!trend) {
       return [];
     }
 
-    try {
-      const items = await this.fetchTrendSourceItems(trend, limit);
-      if (items.length > 0) {
-        return this.trendReferenceCorpusService.annotateSourceItemsWithReferenceIds(
-          items,
-        );
-      }
-    } catch (error: unknown) {
-      this.loggerService.warn('Failed to fetch live trend source items', {
-        error: error instanceof Error ? error.message : String(error),
-        platform: trend.platform,
-        trendId,
-      });
-    }
-
-    return this.trendReferenceCorpusService.annotateSourceItemsWithReferenceIds(
-      this.buildFallbackTrendSourceItems(trend),
-    );
+    return this.trendSourcePreviewService.getAnnotatedSourceItems(trend, limit);
   }
 
-  async getTrendContent(
+  getTrendContent(
     organizationId?: string,
     brandId?: string,
     options: {
@@ -1038,61 +552,16 @@ export class TrendsService {
       refresh?: boolean;
     } = {},
   ): Promise<TrendContentResult> {
-    const limit = options.limit ?? 30;
-    const platform = options.platform;
-    const refresh = options.refresh ?? false;
-    const cacheKey = this.cacheService.generateKey(
-      'trends:content',
-      organizationId || 'global',
-      brandId || 'global',
-      platform || 'all',
-      limit,
+    return this.trendSourcePreviewService.getTrendContent(
+      { brandId, organizationId },
+      options,
+      () =>
+        this.getTrendsWithAccessControl(
+          organizationId,
+          brandId,
+          options.platform,
+        ),
     );
-
-    if (!refresh) {
-      const cached = await this.cacheService.get<TrendContentResult>(cacheKey);
-      if (cached) {
-        return cached;
-      }
-    }
-
-    const result = await this.getTrendsWithAccessControl(
-      organizationId,
-      brandId,
-      platform,
-    );
-
-    const trends = refresh
-      ? await this.precomputeTrendSourcePreview(result.trends, {
-          force: true,
-          limit: this.SOURCE_PREVIEW_LIMIT,
-        })
-      : result.trends;
-
-    const contentItems =
-      await this.trendReferenceCorpusService.annotateSourceItemsWithReferenceIds(
-        this.buildTrendContentItems(trends, limit),
-      );
-    const payload: TrendContentResult = {
-      connectedPlatforms: result.connectedPlatforms,
-      items: contentItems,
-      latestTrendAt: this.getLatestTrendAt(trends),
-      lockedPlatforms: result.lockedPlatforms,
-      totalTrends: trends.filter((trend) =>
-        this.CONTENT_FEED_PLATFORMS.has(trend.platform),
-      ).length,
-    };
-
-    await this.cacheService.set(cacheKey, payload, {
-      tags: [
-        'trends',
-        'trends:content',
-        platform ? `trends:content:${platform}` : 'trends:content:all',
-      ],
-      ttl: this.CONTENT_CACHE_TTL_SECONDS,
-    });
-
-    return payload;
   }
 
   /**
@@ -1211,649 +680,5 @@ export class TrendsService {
    */
   getTrendTimeline(days: 7 | 30 | 90): Promise<TrendTimelineEntry[]> {
     return this.trendVideoService.getTrendTimeline(days);
-  }
-
-  private async fetchTrendSourceItems(
-    trend: TrendEntity,
-    limit: number,
-  ): Promise<TrendSourceItem[]> {
-    switch (trend.platform) {
-      case 'instagram':
-        return this.fetchInstagramSourceItems(trend, limit);
-      case 'tiktok':
-        return this.fetchTikTokSourceItems(trend, limit);
-      case 'twitter':
-        return this.fetchTwitterSourceItems(trend, limit);
-      case 'youtube':
-        return this.fetchYouTubeSourceItems(trend, limit);
-      case 'reddit':
-        return this.fetchRedditSourceItems(trend, limit);
-      default:
-        return [];
-    }
-  }
-
-  private async fetchInstagramSourceItems(
-    trend: TrendEntity,
-    limit: number,
-  ): Promise<TrendSourceItem[]> {
-    const hashtag = this.getTrendSearchTerm(trend);
-    if (!hashtag) {
-      return [];
-    }
-
-    const posts = await this.apifyService.searchInstagramByHashtag(hashtag, {
-      limit,
-    });
-
-    return posts
-      .map((post) => this.mapInstagramPostToSourceItem(post))
-      .filter((item): item is TrendSourceItem => item !== null);
-  }
-
-  private async fetchTikTokSourceItems(
-    trend: TrendEntity,
-    limit: number,
-  ): Promise<TrendSourceItem[]> {
-    const hashtag = this.getTrendSearchTerm(trend);
-    if (!hashtag) {
-      return [];
-    }
-
-    const videos = await this.apifyService.searchTikTokByHashtag(hashtag, {
-      limit,
-    });
-
-    return videos
-      .map((video) => this.mapTikTokVideoToSourceItem(video))
-      .filter((item): item is TrendSourceItem => item !== null);
-  }
-
-  private async fetchTwitterSourceItems(
-    trend: TrendEntity,
-    limit: number,
-  ): Promise<TrendSourceItem[]> {
-    const query = this.getTrendSearchTerm(trend);
-    if (!query) {
-      return [];
-    }
-
-    const tweets = await this.apifyService.searchTwitterTweets(query, {
-      limit,
-    });
-
-    return tweets.map((tweet) => this.mapTweetToSourceItem(tweet));
-  }
-
-  private async fetchYouTubeSourceItems(
-    trend: TrendEntity,
-    limit: number,
-  ): Promise<TrendSourceItem[]> {
-    const query = this.getTrendSearchTerm(trend);
-    if (!query) {
-      return [];
-    }
-
-    const videos = await this.apifyService.searchYouTubeVideos(query, {
-      limit,
-    });
-
-    return videos
-      .map((video) => this.mapYouTubeVideoToSourceItem(video))
-      .filter((item): item is TrendSourceItem => item !== null);
-  }
-
-  private async fetchRedditSourceItems(
-    trend: TrendEntity,
-    limit: number,
-  ): Promise<TrendSourceItem[]> {
-    const query = this.getTrendSearchTerm(trend);
-    if (!query) {
-      return [];
-    }
-
-    const posts = await this.apifyService.searchRedditPosts(query, { limit });
-
-    return posts
-      .map((post) => this.mapRedditPostToSourceItem(post))
-      .filter((item): item is TrendSourceItem => item !== null);
-  }
-
-  private getTrendSearchTerm(trend: TrendEntity): string | null {
-    const primaryHashtag = Array.isArray(trend.metadata?.hashtags)
-      ? trend.metadata.hashtags.find((tag) => !!tag?.trim())
-      : null;
-
-    if (primaryHashtag) {
-      return primaryHashtag.replace(/^#/, '').trim();
-    }
-
-    const topic = trend.topic?.trim();
-    if (!topic) {
-      return null;
-    }
-
-    return topic.replace(/^#/, '').trim();
-  }
-
-  private mapInstagramPostToSourceItem(
-    post: ApifyInstagramPost,
-  ): TrendSourceItem | null {
-    const sourceUrl = post.shortCode
-      ? `https://www.instagram.com/p/${post.shortCode}/`
-      : post.videoUrl || post.imageUrl;
-
-    if (!sourceUrl) {
-      return null;
-    }
-
-    return {
-      authorHandle: post.ownerUsername || undefined,
-      contentType: post.videoUrl ? 'video' : 'image',
-      id: post.id,
-      mediaUrl: post.videoUrl || post.imageUrl,
-      metrics: {
-        comments: post.commentsCount,
-        likes: post.likesCount,
-        views: post.videoViewCount,
-      },
-      platform: 'instagram',
-      publishedAt: post.timestamp || undefined,
-      sourceUrl,
-      text: post.caption,
-      thumbnailUrl: post.imageUrl,
-      title: this.truncateText(post.caption, 100),
-    };
-  }
-
-  private mapTikTokVideoToSourceItem(
-    video: ApifyTikTokVideo,
-  ): TrendSourceItem | null {
-    if (!video.webVideoUrl) {
-      return null;
-    }
-
-    return {
-      authorHandle:
-        video.authorMeta?.name || video.authorMeta?.nickname || undefined,
-      contentType: 'video',
-      id: video.id,
-      mediaUrl: video.webVideoUrl,
-      metrics: {
-        comments: video.commentCount,
-        likes: video.diggCount,
-        shares: video.shareCount,
-        views: video.playCount,
-      },
-      platform: 'tiktok',
-      publishedAt: video.createTime
-        ? new Date(video.createTime * 1000).toISOString()
-        : undefined,
-      sourceUrl: video.webVideoUrl,
-      text: video.desc,
-      thumbnailUrl: video.authorMeta?.avatar || video.musicMeta?.coverUrl,
-      title: this.truncateText(video.desc, 100),
-    };
-  }
-
-  private mapTweetToSourceItem(tweet: ApifyNormalizedTweet): TrendSourceItem {
-    const sourceUrl = tweet.authorUsername
-      ? `https://x.com/${tweet.authorUsername}/status/${tweet.id}`
-      : `https://x.com/i/web/status/${tweet.id}`;
-
-    return {
-      authorHandle: tweet.authorUsername || undefined,
-      contentType: 'tweet',
-      id: tweet.id,
-      metrics: {
-        comments: tweet.metrics?.replies,
-        likes: tweet.metrics?.likes,
-        shares: tweet.metrics?.retweets,
-      },
-      platform: 'twitter',
-      publishedAt: tweet.createdAt?.toISOString(),
-      sourceUrl,
-      text: tweet.text,
-      thumbnailUrl: tweet.authorAvatarUrl,
-      title: this.truncateText(tweet.text, 100),
-    };
-  }
-
-  private mapYouTubeVideoToSourceItem(
-    video: ApifyYouTubeVideo,
-  ): TrendSourceItem | null {
-    if (!video.url) {
-      return null;
-    }
-
-    return {
-      authorHandle: video.channelName || undefined,
-      contentType: 'video',
-      id: video.id,
-      mediaUrl: video.url,
-      metrics: {
-        comments: video.commentCount,
-        likes: video.likeCount,
-        views: video.viewCount,
-      },
-      platform: 'youtube',
-      publishedAt: video.publishedAt || undefined,
-      sourceUrl: video.url,
-      text: video.description,
-      thumbnailUrl: video.thumbnailUrl,
-      title: video.title,
-    };
-  }
-
-  private mapRedditPostToSourceItem(
-    post: ApifyRedditPost,
-  ): TrendSourceItem | null {
-    const sourceUrl = post.permalink
-      ? `https://reddit.com${post.permalink}`
-      : post.url;
-
-    if (!sourceUrl) {
-      return null;
-    }
-
-    return {
-      authorHandle: post.author || undefined,
-      contentType: post.isVideo ? 'video' : 'post',
-      id: post.id,
-      mediaUrl: post.url,
-      metrics: {
-        comments: post.numComments,
-        likes: post.score,
-      },
-      platform: 'reddit',
-      publishedAt: post.createdUtc
-        ? new Date(post.createdUtc * 1000).toISOString()
-        : undefined,
-      sourceUrl,
-      text: post.title,
-      title: post.title,
-    };
-  }
-
-  private buildFallbackTrendSourceItems(trend: TrendEntity): TrendSourceItem[] {
-    const mediaUrl =
-      typeof trend.metadata?.videoUrl === 'string'
-        ? trend.metadata.videoUrl
-        : undefined;
-    const sourceUrls = Array.isArray(trend.metadata?.urls)
-      ? trend.metadata.urls.filter(
-          (url): url is string => typeof url === 'string' && !!url,
-        )
-      : [];
-    const resolvedSourceUrls =
-      sourceUrls.length > 0 ? sourceUrls : mediaUrl ? [mediaUrl] : [];
-
-    return resolvedSourceUrls.map((sourceUrl, index) => ({
-      authorHandle:
-        typeof trend.metadata?.creatorHandle === 'string'
-          ? trend.metadata.creatorHandle
-          : undefined,
-      contentType:
-        trend.platform === 'twitter'
-          ? 'tweet'
-          : trend.metadata?.videoUrl
-            ? 'video'
-            : 'post',
-      id: `${trend.id}-fallback-${index + 1}`,
-      mediaUrl,
-      platform: trend.platform,
-      publishedAt: trend.createdAt?.toISOString(),
-      sourceUrl,
-      text:
-        typeof trend.metadata?.sampleContent === 'string'
-          ? trend.metadata.sampleContent
-          : trend.topic,
-      thumbnailUrl:
-        typeof trend.metadata?.thumbnailUrl === 'string'
-          ? trend.metadata.thumbnailUrl
-          : undefined,
-      title: trend.topic,
-    }));
-  }
-
-  private getStoredTrendSourcePreview(
-    trend: TrendEntity,
-    limit: number = this.SOURCE_PREVIEW_LIMIT,
-  ): TrendSourceItem[] {
-    const cachedItems = trend.metadata?.sourcePreviewCache;
-    if (!Array.isArray(cachedItems)) {
-      return [];
-    }
-
-    return cachedItems
-      .filter((item): item is TrendSourceItem => this.isTrendSourceItem(item))
-      .slice(0, limit);
-  }
-
-  private getStoredTrendSourcePreviewState(
-    trend: TrendEntity,
-    items: TrendSourceItem[],
-  ): 'live' | 'fallback' | 'empty' {
-    const storedState = trend.metadata?.sourcePreviewState;
-    if (
-      storedState === 'live' ||
-      storedState === 'fallback' ||
-      storedState === 'empty'
-    ) {
-      return storedState;
-    }
-
-    return this.getSourcePreviewState(items);
-  }
-
-  private getSourcePreviewState(
-    items: TrendSourceItem[],
-  ): 'live' | 'fallback' | 'empty' {
-    if (items.length === 0) {
-      return 'empty';
-    }
-
-    return items[0]?.id.includes('-fallback') ? 'fallback' : 'live';
-  }
-
-  private isTrendSourceItem(value: unknown): value is TrendSourceItem {
-    if (!value || typeof value !== 'object') {
-      return false;
-    }
-
-    const item = value as Record<string, unknown>;
-    return (
-      typeof item.id === 'string' &&
-      typeof item.platform === 'string' &&
-      typeof item.sourceUrl === 'string' &&
-      typeof item.contentType === 'string'
-    );
-  }
-
-  private async persistTrendSourcePreview(
-    trend: TrendEntity,
-    items: TrendSourceItem[],
-  ): Promise<TrendEntity> {
-    const sourcePreviewState = this.getSourcePreviewState(items);
-    const metadata = {
-      ...trend.metadata,
-      sourcePreviewCache: items,
-      sourcePreviewCachedAt: new Date().toISOString(),
-      sourcePreviewState,
-    };
-
-    const existingDoc = await this.prisma.trend.findFirst({
-      where: { id: String(trend.id), isDeleted: false } as never,
-    });
-    if (existingDoc) {
-      const existingData =
-        (existingDoc.data as unknown as Record<string, unknown>) ?? {};
-      await this.prisma.trend.update({
-        data: { data: { ...existingData, metadata } as never },
-        where: { id: String(trend.id) },
-      });
-    }
-
-    return new TrendEntity({
-      ...trend,
-      metadata,
-    });
-  }
-
-  private async precomputeTrendSourcePreview(
-    trends: TrendEntity[],
-    options: {
-      force?: boolean;
-      limit?: number;
-    } = {},
-  ): Promise<TrendEntity[]> {
-    const limit = options.limit ?? this.SOURCE_PREVIEW_LIMIT;
-
-    return await Promise.all(
-      trends.map(async (trend) => {
-        if (!this.CONTENT_FEED_PLATFORMS.has(trend.platform)) {
-          return trend;
-        }
-
-        const cachedItems = this.getStoredTrendSourcePreview(trend, limit);
-        if (!options.force && cachedItems.length > 0) {
-          return trend;
-        }
-
-        const resolvedItems = await this.resolveTrendSourcePreview(
-          trend,
-          undefined,
-          {
-            allowLiveFetch: true,
-            ignoreCached: options.force === true,
-            limit,
-          },
-        );
-
-        return this.persistTrendSourcePreview(trend, resolvedItems);
-      }),
-    );
-  }
-
-  private getTrendContentCacheKey(item: TrendContentItem): string {
-    return `${item.platform}:${item.sourceUrl || item.id}`;
-  }
-
-  private getTrendContentEngagementScore(item: TrendContentItem): number {
-    return (
-      (item.metrics?.views || 0) +
-      (item.metrics?.likes || 0) +
-      (item.metrics?.comments || 0) +
-      (item.metrics?.shares || 0)
-    );
-  }
-
-  private compareTrendContentItems(
-    left: TrendContentItem,
-    right: TrendContentItem,
-  ): number {
-    const liveStateDelta =
-      Number(right.sourcePreviewState === 'live') -
-      Number(left.sourcePreviewState === 'live');
-    if (liveStateDelta !== 0) {
-      return liveStateDelta;
-    }
-
-    const viralityDelta = right.trendViralityScore - left.trendViralityScore;
-    if (viralityDelta !== 0) {
-      return viralityDelta;
-    }
-
-    const engagementDelta =
-      this.getTrendContentEngagementScore(right) -
-      this.getTrendContentEngagementScore(left);
-    if (engagementDelta !== 0) {
-      return engagementDelta;
-    }
-
-    return (
-      new Date(right.publishedAt || 0).getTime() -
-      new Date(left.publishedAt || 0).getTime()
-    );
-  }
-
-  private buildTrendContentItems(
-    trends: TrendEntity[],
-    limit: number,
-  ): TrendContentItem[] {
-    const dedupedItems = new Map<string, TrendContentItem>();
-
-    for (const trend of trends) {
-      if (!this.CONTENT_FEED_PLATFORMS.has(trend.platform)) {
-        continue;
-      }
-
-      const storedItems = this.getStoredTrendSourcePreview(
-        trend,
-        this.SOURCE_PREVIEW_LIMIT,
-      );
-      const sourceItems =
-        storedItems.length > 0
-          ? storedItems
-          : this.buildFallbackTrendSourceItems(trend);
-
-      for (const sourceItem of sourceItems) {
-        const contentItem: TrendContentItem = {
-          ...sourceItem,
-          contentRank: 0,
-          matchedTrends: [trend.topic],
-          requiresAuth: trend.requiresAuth,
-          sourcePreviewState: this.getSourcePreviewState([sourceItem]),
-          trendId: String(trend.id),
-          trendMentions: trend.mentions,
-          trendTopic: trend.topic,
-          trendViralityScore: trend.viralityScore,
-        };
-        const dedupeKey = this.getTrendContentCacheKey(contentItem);
-        const existing = dedupedItems.get(dedupeKey);
-
-        if (!existing) {
-          dedupedItems.set(dedupeKey, contentItem);
-          continue;
-        }
-
-        const matchedTrends = Array.from(
-          new Set([...existing.matchedTrends, trend.topic]),
-        );
-        const preferred =
-          this.compareTrendContentItems(contentItem, existing) > 0
-            ? existing
-            : contentItem;
-
-        dedupedItems.set(dedupeKey, {
-          ...preferred,
-          matchedTrends,
-        });
-      }
-    }
-
-    return Array.from(dedupedItems.values())
-      .sort((left, right) => this.compareTrendContentItems(left, right))
-      .slice(0, limit)
-      .map((item, index) => ({
-        ...item,
-        contentRank: index + 1,
-      }));
-  }
-
-  private getLatestTrendAt(trends: TrendEntity[]): string | undefined {
-    const timestamps = trends
-      .map((trend) =>
-        trend.createdAt instanceof Date
-          ? trend.createdAt.getTime()
-          : trend.createdAt
-            ? new Date(trend.createdAt).getTime()
-            : 0,
-      )
-      .filter((value) => value > 0);
-
-    if (timestamps.length === 0) {
-      return undefined;
-    }
-
-    return new Date(Math.max(...timestamps)).toISOString();
-  }
-
-  private truncateText(
-    value?: string,
-    maxLength: number = 100,
-  ): string | undefined {
-    if (!value) {
-      return undefined;
-    }
-
-    return value.length > maxLength
-      ? `${value.slice(0, maxLength - 1)}…`
-      : value;
-  }
-
-  private async buildTrendDiscoveryItem(
-    trend: TrendEntity,
-    organizationId?: string,
-  ): Promise<TrendDiscoveryItem> {
-    const sourcePreview = await this.resolveTrendSourcePreview(
-      trend,
-      organizationId,
-      {
-        allowLiveFetch: false,
-        limit: 3,
-      },
-    );
-    const sourcePreviewState = this.getStoredTrendSourcePreviewState(
-      trend,
-      sourcePreview,
-    );
-
-    return {
-      ...trend,
-      id: String(trend.id),
-      createdAt:
-        trend.createdAt instanceof Date
-          ? trend.createdAt
-          : trend.createdAt
-            ? new Date(trend.createdAt)
-            : undefined,
-      expiresAt:
-        trend.expiresAt instanceof Date
-          ? trend.expiresAt
-          : new Date(trend.expiresAt),
-      sourcePreview,
-      sourcePreviewState,
-      sourcePreviewTotal: sourcePreview.length,
-      updatedAt:
-        trend.updatedAt instanceof Date
-          ? trend.updatedAt
-          : trend.updatedAt
-            ? new Date(trend.updatedAt)
-            : undefined,
-    };
-  }
-
-  private async resolveTrendSourcePreview(
-    trend: TrendEntity,
-    organizationId?: string,
-    options: {
-      allowLiveFetch?: boolean;
-      ignoreCached?: boolean;
-      limit?: number;
-    } = {},
-  ): Promise<TrendSourceItem[]> {
-    const limit = options.limit ?? this.SOURCE_PREVIEW_LIMIT;
-    const cachedItems =
-      options.ignoreCached === true
-        ? []
-        : this.getStoredTrendSourcePreview(trend, limit);
-    if (cachedItems.length > 0) {
-      return cachedItems;
-    }
-
-    const fallbackItems = this.buildFallbackTrendSourceItems(trend).slice(
-      0,
-      limit,
-    );
-    if (!options.allowLiveFetch) {
-      return fallbackItems;
-    }
-
-    try {
-      const items = await this.fetchTrendSourceItems(trend, limit);
-      if (items.length > 0) {
-        return items;
-      }
-    } catch (error: unknown) {
-      this.loggerService.warn('Failed to build embedded trend source preview', {
-        error: error instanceof Error ? error.message : String(error),
-        organizationId,
-        platform: trend.platform,
-        trendId: trend.id,
-      });
-    }
-
-    return fallbackItems;
   }
 }
